+++ metadata/Cargo.toml
@@ -9,6 +9,6 @@ edition = "2018"
 codec = { package = "parity-scale-codec", version = "1.3.5", default-features = false }
 
 # Substrate
-frame-metadata = { version = "12", package = "frame-metadata", default-features = false }
+frame-metadata = { version = "12", package = "frame-metadata" }
 
 substrate-airgapped = { version = "0.1.0", path = "../substrate-airgapped" }
