+++ substrate-airgapped/Cargo.toml
@@ -15,7 +15,7 @@ sp-core = { version = "2.0.0", default-features = false }
 [dev-dependencies]
 # Currently these are just to get the example going
 sp-keyring = { version = "2.0.0", package = "sp-keyring", default-features = false }
-sp-version = { version = "2.0.0", package = "sp-version", default-features = false }
+sp-version = { version = "2.0.0", package = "sp-version" }
 hex = { version = "0.4.2" }
 serde = { version = "1.0.115", features = ["derive"] }
 serde_json = "1.0.57"
