+++ substrate-airgapped/examples/call_index_from_metadata.rs
@@ -1,7 +1,7 @@
 use ::core::convert::TryInto;
 use codec::Decode;
 use frame_metadata::RuntimeMetadataPrefixed;
-use substrate_airgapped::{balances::Transfer, GenericCall};
+use substrate_airgapped::{balances::Transfer, GenericCall, KusamaRuntime};
 
 use metadata::Metadata;
 use serde::{Deserialize, Serialize};
@@ -9,19 +9,27 @@ use sp_keyring::AccountKeyring;
 use sp_runtime::DeserializeOwned;
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
-	let metadata_bytes = rpc_to_local_node::<(), String>("chain_getBlockHash", vec![])
+	let metadata_bytes = rpc_to_local_node::<(), String>("state_getMetadata", vec![])
 		.and_then(|rpc_res| Ok(hex::decode(rpc_res.result)?))?;
 
 	let metadata_prefixed: RuntimeMetadataPrefixed = Decode::decode(&mut &metadata_bytes[..])?;
 	let metadata: Metadata = metadata_prefixed.try_into()?;
 
-	let args = Transfer { to: AccountKeyring::Bob.to_account_id().into(), amount: 123_456_789 };
-	let call_index = metadata.find_call_index(args)?;
+	let args: Transfer<KusamaRuntime> = Transfer { to: AccountKeyring::Bob.to_account_id().into(), amount: 123_456_789 };
+	// Use the `CallMethod` to fetch the `CallIndex` of the dispatchable
+	let call_index = metadata.find_call_index(&args)?;
+	println!("Call index for balances::Transfer: {:#?}", call_index);
+
+	// You can then use the returned call index to construct a `GenericCall` - the type needed to
+	// construct a `UncheckedExtrinsic`
 	let transfer_call = GenericCall::new(call_index, args);
+	println!("SCALE Encode-able balances::Transfer call: {:#?}", transfer_call);
 
 	Ok(())
 }
 
+// These are redundant utils as they are also used in the other example.
+// There should be a better solution.
 /// RPC response JSON object
 #[derive(Serialize, Deserialize, Debug)]
 pub struct RpcRes<T> {
